adaptor plugin mechanism phase1
